# src/router/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telicent ACCESS",
    "version": "0.11.2",
    "description": "Telicent ACCESS API",
    "contact": {
      "name": "Telicent",
      "url": "https://telicent.io",
      "email": "info@telicent.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8091"
    }
  ],
  "paths": {
...
    "/hierarchies": {
      "get": {
        "summary": "Get all hierarchies",
        "tags": [
          "Hierarchies"
        ],
        "description": "Returns all hierarchies",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all hierarchies",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Hierarchy"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Hierarchies not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HierarchyNotFound"
                }
              }
            }
          }
        }
      }
    },
    "/hierarchies/{hierarchyId}": {
      "get": {
        "summary": "Get hierarchy with id :hierarchyId",
        "tags": [
          "Hierarchies"
        ],
        "description": "Returns hierarchy with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "hierarchyId",
            "description": "Identifier of requested hierarchy"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved hierarchy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hierarchy"
                }
              }
            }
          },
          "404": {
            "description": "Hierarchy not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HierarchyNotFound"
                }
              }
            }
          }
        }
      }
    },
    "/hierarchies/lookup/{name}": {
      "get": {
        "summary": "Look up hierarchy by name",
        "tags": [
          "Hierarchies"
        ],
        "description": "Looks up hierarchy by name, using the dataAttributeName or userAttributeName property of the hierarchy attribute",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "description": "Name of the the hierarchy to be looked up"
          },
          {
            "in": "query",
            "name": "isUserAttribute",
            "schema": {
              "type": "boolean",
              "required": false,
              "default": false,
              "description": "isUserAttribute being set to true tells ACCESS to look up the hierarchy using the userAttributeName property rather than dataAttributeName"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully looked up hierarchy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hierarchy"
                }
              }
            }
          },
          "404": {
            "description": "Hierarchy not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HierarchyNotFound"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Hierarchy": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "example": "64d25a470a97a540165757da"
          },
          "name": {
            "type": "string",
            "example": "classification"
          },
          "tiers": {
            "type": "array",
            "example": [
              "O",
              "OS",
              "S",
              "TS"
            ]
          },
          "readOnly": {
            "type": "boolean"
          }
        }
      },
      "HierarchyNotFound": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "hierarchy not found"
          }
        }
      },
...
    }
  },
...
}